+++ packages/onboarding/src/steps/Carousel/types.ts
@@ -1,5 +1,5 @@
 import { z } from "zod";
-import { ButtonSectionSchema, CustomPayloadSchema } from "../common.types";
+import { BaseStepTypeSchema } from "../common.types";
 
 export const CarouselScreenSchema = z.object({
   mediaUrl: z.string(),
@@ -11,16 +11,9 @@ export const CarouselStepPayloadSchema = z.object({
   screens: z.array(CarouselScreenSchema),
 });
 
-export const CarouselStepTypeSchema = z.object({
-  id: z.string(),
+export const CarouselStepTypeSchema = BaseStepTypeSchema.extend({
   type: z.literal("Carousel"),
-  name: z.string(),
-  displayProgressHeader: z.boolean(),
   payload: CarouselStepPayloadSchema,
-  customPayload: CustomPayloadSchema,
-  continueButtonLabel: z.string().optional().default("Continue"),
-  buttonSection: ButtonSectionSchema.optional(),
-  figmaUrl: z.string().nullish(),
 });
 
 export type CarouselStepType = z.infer<typeof CarouselStepTypeSchema>;