+++ packages/onboarding/src/steps/Ratings/types.ts
@@ -1,5 +1,5 @@
 import z from "zod";
-import { ButtonSectionSchema, CustomPayloadSchema, SocialProofSchema } from "../common.types";
+import { BaseStepTypeSchema, SocialProofSchema } from "../common.types";
 
 export const RatingsStepPayloadSchema = z.object({
   title: z.string(),
@@ -8,16 +8,9 @@ export const RatingsStepPayloadSchema = z.object({
   rateTheAppButtonLabel: z.string().optional().default("Rate the app"),
 });
 
-export const RatingsStepTypeSchema = z.object({
-  id: z.string(),
+export const RatingsStepTypeSchema = BaseStepTypeSchema.extend({
   type: z.literal("Ratings"),
-  name: z.string(),
-  displayProgressHeader: z.boolean(),
   payload: RatingsStepPayloadSchema,
-  customPayload: CustomPayloadSchema,
-  continueButtonLabel: z.string().optional().default("Continue"),
-  buttonSection: ButtonSectionSchema.optional(),
-  figmaUrl: z.string().nullish(),
 });
 
 export type RatingsStepType = z.infer<typeof RatingsStepTypeSchema>;